Added database, models, routes and wired them all
Integrated router with a file module that fetches every route in
the routes folder and mount it as a new route.

Integrated database to be used inside routes middlewares as models
loaded from any files present inside models folder

Tried an approach in case postgres database is absent to use sqlite
in-memory database to implement test cases and other stuff

Added a wrong README, will get this right in next commit.